+++ script/ability/effect.ts
@@ -73,7 +73,7 @@ class Effect {
     return tooltip;
   }
 
-  init(bonuses: any) {
+  init(bonuses: any, caster: any) {
     if (!bonuses) bonuses = {};
     Object.entries(this).forEach(([key, value]) => {
       if (typeof value === "number") {
@@ -90,7 +90,15 @@ class Effect {
         });
       }
     });
-
+    console.log(caster);
+    if (this.inflict && caster) {
+      if (this.inflict.healingFlat) {
+        this.inflict.healingFlat = Math.round(this.inflict.healingFlat * caster.allModifiers.healPowerP);
+      }
+      if (this.inflict.healingPercent) {
+        this.inflict.healingPercent = +(this.inflict.healingPercent * caster.allModifiers.healPowerP).toFixed(3);
+      }
+    }
     return this;
   }
 }