+++ script/compiled/ability/ability.js
@@ -4,6 +4,7 @@ class Ability {
     mpCost;
     hpCost;
     type;
+    skillType; // Either melee or ranged
     isSpell;
     cooldown;
     onCooldown;
@@ -23,6 +24,7 @@ class Ability {
         this.mpCost = ability.mpCost ?? 0;
         this.hpCost = ability.hpCost ?? 0;
         this.type = ability.type;
+        this.skillType = ability.skillType ?? "";
         this.isSpell = ability.isSpell ?? false;
         this.weight = ability.weight ?? 1;
         this.isAOE = ability.isAOE ?? false;
@@ -205,6 +207,10 @@ class Ability {
                     this[key] = { ...updateObject(key, value, holder.allModifiers[id]) };
                 }
             });
+            if (this.healFlat) {
+                this.healFlat = Math.round(this.healFlat * holder.allModifiers.healPowerP);
+                this.healPercent = Math.round(this.healPercent * holder.allModifiers.healPowerP);
+            }
         };
     }
     tooltip(options) {
@@ -266,7 +272,7 @@ class Ability {
             this.effectsToEnemy.forEach((effect) => {
                 if (options?.owner) {
                     const displayEffect = new Effect(effect);
-                    displayEffect.init(options?.owner?.allModifiers?.["ability_" + this.id]?.["effect_" + effect.id]);
+                    displayEffect.init(options?.owner?.allModifiers?.["ability_" + this.id]?.["effect_" + effect.id], options?.owner);
                     tooltip += displayEffect.tooltip({ container: true });
                 }
                 else {
@@ -279,7 +285,7 @@ class Ability {
             this.effectsToSelf.forEach((effect) => {
                 if (options?.owner) {
                     const displayEffect = new Effect(effect);
-                    displayEffect.init(options?.owner?.allModifiers?.["ability_" + this.id]?.["effect_" + effect.id]);
+                    displayEffect.init(options?.owner?.allModifiers?.["ability_" + this.id]?.["effect_" + effect.id], options?.owner);
                     tooltip += displayEffect.tooltip({ container: true });
                 }
                 else {