+++ script/dev/dev.ts
@@ -16,7 +16,7 @@ const DEVTOOLS: DevTools = {
 
 if (localStorage.getItem("devtools") === "true") DEVTOOLS.ENABLED = true;
 
-const gameVersion = (0.12).toFixed(2);
+const gameVersion = (0.13).toFixed(2);
 const gameVersionText = (ver: string) => {
   if (parseFloat(ver) < 0.1) return "Old";
   return `${parseInt(ver)}.${(parseFloat(ver) * 10) % 10}`;