+++ script/dev/new_game.ts
@@ -91,6 +91,11 @@ const characterCreationSections: any = [
     type: "input",
   },
   {
+    id: "class",
+    type: "aspect",
+    values: [classManager.get("warrior"), classManager.get("rogue"), classManager.get("mage"), classManager.get("paladin")],
+  },
+  {
     id: "starting_aspect",
     type: "aspect",
     values: [
@@ -163,12 +168,16 @@ function createSectionValue(section: any) {
         const gridItem = document.createElement("div");
         gridItem.classList.add("choice");
         const gridItemName = document.createElement("p");
-        if (player.starting_aspect === value.id) {
+        if (player.starting_aspect === value.id || player.class.id === value.id) {
           gridItem.classList.add("selected");
         }
 
         gridItem.onclick = () => {
-          player.starting_aspect = value.id;
+          if (value.type === "class") {
+            player.class = new CharClass(value);
+          } else {
+            player.starting_aspect = value.id;
+          }
           grid.querySelectorAll(".choice").forEach((item) => {
             item.classList.remove("selected");
           });
@@ -238,7 +247,6 @@ function aspectTooltip(aspect: any) {
   if (aspect.modifiers) {
     tooltip += `<f>1rem<f><c>white<c>${game.getLocalizedString("modifiers")}:<c>white<c>\n`;
     Object.entries(aspect.modifiers).forEach(([key, value]: [string, any]) => {
-      console.log(key, value);
       tooltip += effectSyntax(key, value);
     });
   }
@@ -247,7 +255,7 @@ function aspectTooltip(aspect: any) {
     aspect.items.forEach((item: any) => {
       tooltip += `<f>1rem<f><c>white<c>${item.amount}x <c>gold<c>${game.getLocalizedString(item.item)}<c>white<c>\n`;
     });
-  } else {
+  } else if (aspect?.type !== "class") {
     tooltip += `<f>1rem<f><c>white<c>${game.getLocalizedString("no_items")}<c>white<c>\n`;
   }
   return tooltip;